% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Velo.R
\name{calcul_distribution_semaine}
\alias{calcul_distribution_semaine}
\title{Calculer la distribution hebdomadaire des trajets}
\usage{
calcul_distribution_semaine(trajet)
}
\arguments{
\item{trajet}{Un data frame contenant les données de trajets.}
}
\value{
Un data frame avec le jour de la semaine et le nombre
total de trajets.
}
\description{
Cette fonction calcule le nombre total de trajets
par jour de la semaine.
}
